Add a contributor list script
